add 2 chart for admin analytic
